Move runtime-templates.ts out of the vendored genql tree
src/generate/genql should contain only the fork of @genql/cli (plus its
provenance metadata); runtime-templates.ts is Twenty's bundling
mechanism, not upstream code, so it now lives beside the tree in
src/generate. The ?raw imports and the tsx disk-read fallback point
into genql/runtime, and the provenance ledger drops its entry.
